BQ_scd2_processing/requirements.txt
pandas
google-cloud-bigquery
google-cloud-storage